end of day
updated AdvancedPopupPage to have an constructor made to enable getting login credentials
added a TestPage that takes a list of views and simply displays them
created a extended "entry" which hold all that should be needed to have an entry field with hide/show text capability. inreality it is a stacklayout, containing multiple view elements to make it work. currently it crashes when trying to created the binding, for some unexplained reason.