Finalized PasswordEntry
Removed the trigger action as it is now part of the PasswordEntry's Clicked event, in doing so i moved the resources over to the elements folder.
Fine tuned the layout generated for the pop up when creating a login page. Might add the possibility to add a image on top of the pop up

Now to update the Dental Project to make use of this new PasswordEntry